# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "ProjectsController_listProjects",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["client", "internal"]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["draft", "active", "on_hold", "completed", "archived"]
            }
          },
          {
            "name": "phase",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "intake",
                "requirements",
                "clarification",
                "baseline",
                "architecture",
                "delivery",
                "handoff",
                "closed"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "critical"]
            }
          },
          {
            "name": "visibility",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["private", "organization"]
            }
          },
          {
            "name": "clientId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "string",
                  "const": "true"
                },
                {
                  "type": "string",
                  "const": "false"
                },
                {
                  "type": "string",
                  "const": "1"
                },
                {
                  "type": "string",
                  "const": "0"
                }
              ]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "updated_desc",
              "type": "string",
              "enum": ["updated_desc", "updated_asc", "name_asc", "name_desc"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A filtered, cursor-paginated project list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed. Error code: VALIDATION_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed."
                }
              }
            }
          },
          "401": {
            "description": "No valid, active session was supplied. Error code: AUTHENTICATION_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "No valid, active session was supplied."
                }
              }
            }
          },
          "403": {
            "description": "The authenticated account is inactive. Error code: ACCESS_DENIED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "ACCESS_DENIED",
                  "message": "The authenticated account is inactive."
                }
              }
            }
          },
          "500": {
            "description": "An unexpected internal error occurred. Error code: INTERNAL_SERVER_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 500,
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "Internal server error."
                }
              }
            }
          }
        },
        "summary": "Search projects visible to the caller in the current organization",
        "tags": ["projects"]
      },
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed. Error code: VALIDATION_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed."
                }
              }
            }
          },
          "401": {
            "description": "No valid, active session was supplied. Error code: AUTHENTICATION_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "No valid, active session was supplied."
                }
              }
            }
          },
          "403": {
            "description": "The authenticated account is inactive or lacks organization-admin privileges. Error code: ACCESS_DENIED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "ACCESS_DENIED",
                  "message": "The authenticated account is inactive or lacks organization-admin privileges."
                }
              }
            }
          },
          "500": {
            "description": "An unexpected internal error occurred. Error code: INTERNAL_SERVER_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 500,
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "Internal server error."
                }
              }
            }
          }
        },
        "summary": "Create a project in the current organization",
        "tags": ["projects"]
      }
    },
...
    "/api/v1/projects/{projectId}/archive": {
      "post": {
        "operationId": "ProjectsController_archiveProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The archived project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed. Error code: VALIDATION_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed."
                }
              }
            }
          },
          "401": {
            "description": "No valid, active session was supplied. Error code: AUTHENTICATION_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "No valid, active session was supplied."
                }
              }
            }
          },
          "403": {
            "description": "The caller cannot archive this project. Error code: ACCESS_DENIED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "ACCESS_DENIED",
                  "message": "The caller cannot archive this project."
                }
              }
            }
          },
          "404": {
            "description": "Project not found. Error code: RESOURCE_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "Project not found."
                }
              }
            }
          },
          "500": {
            "description": "An unexpected internal error occurred. Error code: INTERNAL_SERVER_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 500,
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "Internal server error."
                }
              }
            }
          }
        },
        "summary": "Archive a project in the current organization",
        "tags": ["projects"]
      }
    },
    "/api/v1/projects/{projectId}/restore": {
      "post": {
        "operationId": "ProjectsController_restoreProject",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The restored project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed. Error code: VALIDATION_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "code": "VALIDATION_ERROR",
                  "message": "Request validation failed."
                }
              }
            }
          },
          "401": {
            "description": "No valid, active session was supplied. Error code: AUTHENTICATION_REQUIRED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "code": "AUTHENTICATION_REQUIRED",
                  "message": "No valid, active session was supplied."
                }
              }
            }
          },
          "403": {
            "description": "The caller cannot restore this project. Error code: ACCESS_DENIED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "code": "ACCESS_DENIED",
                  "message": "The caller cannot restore this project."
                }
              }
            }
          },
          "404": {
            "description": "Project not found. Error code: RESOURCE_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "Project not found."
                }
              }
            }
          },
          "500": {
            "description": "An unexpected internal error occurred. Error code: INTERNAL_SERVER_ERROR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 500,
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "Internal server error."
                }
              }
            }
          }
        },
        "summary": "Restore an archived project to active",
        "tags": ["projects"]
      }
    },
...
  },
  "info": {
    "title": "AtlasHQ API",
    "description": "AtlasHQ REST API contract. Operation IDs use Controller_method.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z0-9_]+$"
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "details": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "minItems": 1,
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      }
                    ]
                  }
                },
                "message": {
                  "type": "string",
                  "minLength": 1
                },
                "code": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": ["path", "message", "code"],
              "additionalProperties": false
            }
          },
          "correlationId": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["statusCode", "code", "message"],
        "additionalProperties": false
      },
...
    }
  }
}